+++ sdk/containerservice/azure-mgmt-containerservicefleet/azure/mgmt/containerservicefleet/_client.py
@@ -70,9 +70,9 @@ class ContainerServiceFleetMgmtClient:  # pylint: disable=too-many-instance-attr
     :keyword cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
      None.
     :paramtype cloud_setting: ~azure.core.AzureClouds
-    :keyword api_version: The API version to use for this operation. Default value is
-     "2025-08-01-preview". Note that overriding this default value may result in unsupported
-     behavior.
+    :keyword api_version: The API version to use for this operation. Known values are
+     "2025-08-01-preview" and None. Default value is "2025-08-01-preview". Note that overriding this
+     default value may result in unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
      Retry-After header is present.