+++ sdk/containerservice/azure-mgmt-containerservicefleet/azure/mgmt/containerservicefleet/_configuration.py
@@ -33,9 +33,9 @@ class ContainerServiceFleetMgmtClientConfiguration:  # pylint: disable=too-many-
     :param cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
      None.
     :type cloud_setting: ~azure.core.AzureClouds
-    :keyword api_version: The API version to use for this operation. Default value is
-     "2025-08-01-preview". Note that overriding this default value may result in unsupported
-     behavior.
+    :keyword api_version: The API version to use for this operation. Known values are
+     "2025-08-01-preview" and None. Default value is "2025-08-01-preview". Note that overriding this
+     default value may result in unsupported behavior.
     :paramtype api_version: str
     """